Names of directories changed using standard
