--- WebSite/api/award_service.py
@@ -1,5 +1,0 @@
-from WebSite.models import Award
-
-
-def get_awards():
-	return Award.objects.all()
