+++ acm_web_site/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for ACMWebSite project.
+WSGI config for acm_web_site project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "ACMWebSite.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "acm_web_site.settings")
 
 application = get_wsgi_application()
