+++ web_site/api/award_service.py
@@ -1,0 +1,5 @@
+from web_site.models import Award
+
+
+def get_awards():
+	return Award.objects.all()
