+++ web_site/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class WebsiteConfig(AppConfig):
-    name = 'WebSite'
+    name = 'web_site'
