+++ web_site/serializer/activity_serializer.py
@@ -8,9 +8,9 @@ from django.utils.http import urlsafe_base64_decode
 from django.contrib.auth.tokens import default_token_generator
 from rest_framework import serializers
 
-from ACMWebSite.settings import MEDIA_ROOT
-from WebSite.models import Member
-from WebSite.models import Activity
+from acm_web_site.settings import MEDIA_ROOT
+from web_site.models import Member
+from web_site.models import Activity
 
 
 class ActivitySerializer(serializers.ModelSerializer):
