+++ web_site/serializer/gallery_serializer.py
@@ -2,17 +2,12 @@
 from __future__ import unicode_literals
 import base64
 
-from ACMWebSite.settings import MEDIA_ROOT
 from rest_framework import serializers
-from WebSite.models import Award
+from acm_web_site.settings import MEDIA_ROOT
 
 
-class AwardSerializer(serializers.ModelSerializer):
+class GallerySerializer(serializers.Serializer):
     picture = serializers.SerializerMethodField()
-
-    class Meta:
-        model = Award
-        fields = '__all__'
 
     def get_picture(self, obj):
         prefix = '/'.join(MEDIA_ROOT.split('/')[:-1])
