+++ web_site/serializer/member_serializer.py
@@ -2,10 +2,10 @@
 from __future__ import unicode_literals
 import base64
 
-from ACMWebSite.settings import MEDIA_ROOT
+from acm_web_site.settings import MEDIA_ROOT
 
 from rest_framework import serializers
-from WebSite.models import Member, POSITION_CHOICES
+from web_site.models import Member, POSITION_CHOICES
 
 
 class MemberSerializer(serializers.ModelSerializer):
