+++ web_site/serializer/team_serializer.py
@@ -3,20 +3,21 @@ from __future__ import unicode_literals
 import base64
 
 from rest_framework import serializers
-from WebSite.models import Project
-from ACMWebSite.settings import MEDIA_ROOT
+
+from acm_web_site.settings import MEDIA_ROOT
+from web_site.models import Team
 
 
-class ProjectSerializer(serializers.ModelSerializer):
-    poster = serializers.SerializerMethodField()
+class TeamSerializer(serializers.ModelSerializer):
+    picture = serializers.SerializerMethodField(required=False)
 
     class Meta:
-        model = Project
+        model = Team
         fields = '__all__'
 
-    def get_poster(self, obj):
+    def get_picture(self, obj):
         prefix = '/'.join(MEDIA_ROOT.split('/')[:-1])
-        complete_path = prefix + obj.poster
+        complete_path = prefix + obj.picture
         with open(complete_path, "rb") as image_file:
             str = base64.b64encode(image_file.read())
         return str
